refactor: JdbcRepository로 의존성 변경-v1

+++ ch14/src/main/java/kr/co/ordermanagement/presentation/dto/OrderResponse.java
@@ -11,7 +11,7 @@ public class OrderResponse {
 
   private Long id;
   private List<OrderedProduct> orderedProducts;
-  private Double totalPrice;
+  private Integer totalPrice;
   private OrderStatus state;
 
   public OrderResponse(Order order) {
